+++ src/App.js
@@ -1,14 +1,24 @@
 import React from 'react';
-import { ThemeProvider } from 'styled-components';
-import Content from './templates/Content';
+import styled, { ThemeProvider } from 'styled-components';
+import Footer from './components/Footer';
+import Header from './components/Header';
+import Homepage from './templates/Homepage';
 import theme from './utils/scss/theme';
 
-function App() {
-  const page = 'homepage';
+const Wrapper = styled.div`
+  * {
+    box-sizing: border-box;
+  }
+`;
 
+function App() {
   return (
     <ThemeProvider theme={theme}>
-      <Content page={page} />
+      <Wrapper>
+        <Header />
+        <Homepage />
+        <Footer />
+      </Wrapper>
     </ThemeProvider>
   );
 }
