+++ src/components/Footer/Footer.test.js
@@ -2,8 +2,9 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import Footer from '.';
 
-test('footer logo image loads', async () => {
-  const { getByAltText } = await render(<Footer />);
-  const logoImage = getByAltText('Deco Choice');
+test('footer logo image loads', () => {
+  render(<Footer />);
+  // eslint-disable-next-line no-restricted-globals
+  const logoImage = screen.getByAltText('Deco Choice');
   expect(logoImage).toHaveAttribute('src', 'logo.png');
 });
